requirements.txt
altair==5.5.0
anyio==4.9.0
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filetype==1.2.0
fonttools==4.48.1
fqdn==1.5.1
future==0.18.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
igraph==0.11.4
ipykernel==6.29.5
ipython==8.34.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.3.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
libmagic==1.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mistune==3.1.3
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.7
plotly==5.24.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-magic-bin==0.4.14
python-telegram-bot==13.15
pytz==2024.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.31.0
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.0.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.3.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
vboxapi==1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xgboost==3.0.0
xyzservices==2023.10.1
zipp==3.20.1
